requirements.txt
# Core Dependencies
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
beautifulsoup4==4.12.2

# Machine Learning & Deep Learning
torch==2.0.1
transformers==4.35.0
sentence-transformers==2.2.2
scikit-learn==1.3.0

# NLP Processing
spacy==3.6.1
nltk==3.8.1

# Graph Analysis & Community Detection
networkx==3.1
python-louvain==0.16

# Bio/Medical APIs
biopython==1.81
arxiv==1.4.8

# PDF & Text Processing
PyMuPDF==1.23.8
lxml==4.9.3

# Visualization & Web App
streamlit==1.28.0
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2
wordcloud==1.9.2
pyvis==0.3.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1